switched impulse functions and added slightly more functionality for double jump

+++ game/Player.h
@@ -309,7 +309,7 @@ public:
 	//////////////////////////////////////////////////
 	
 	//4 movement mods stuff
-	int						fuel;
+	float					fuel;
 	bool					jetpacking;
 	int						numJumps;
